flask-backend/requirements.txt
SQLAlchemy
Flask
Flask-SQLAlchemy
python-dotenv
flask_jwt_extended
argon2-cffi
flask-restplus
